requirements.txt
flask==2.3.3
flask_sqlalchemy==3.1.1
flask_cors==4.0.0
requests==2.31.0
lxml==4.9.3
psycopg2-binary==2.9.9
neo4j==5.17.0
